chore: commit pre-commit auto-refactor (Link, OuslyImage internal extraction)
